Minor adjustments to the worktree structure.
