--- gix-worktree/tests/worktree/fs/mod.rs
@@ -1,1 +1,0 @@
-mod cache;
