--- gix-worktree/tests/worktree/index/mod.rs
@@ -1,2 +1,0 @@
-mod checkout;
-mod status;
